fixed so that user can log in and also can shift between sign in and signup views
