--- MessengerApp/ViewModels/SignUpViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  SignUpViewViewModel.swift
-//  MessengerApp
-//
-//  Created by Juhee Kang Johansson on 2024-04-04.
-//
-
-import Foundation
